requirements.txt
# Core ML
ultralytics==8.2.0
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python>=4.9.0.80
Pillow>=10.3.0
albumentations>=1.4.0

# API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0
python-multipart>=0.0.9
pydantic>=2.7.0

# Data
numpy>=1.26.4
pandas>=2.2.2
scikit-learn>=1.4.2

# Experiment Tracking
wandb>=0.16.6

# Utilities
tqdm>=4.66.4
PyYAML>=6.0.1
python-dotenv>=1.0.1
requests>=2.31.0

# Visualization
matplotlib>=3.8.4
seaborn>=0.13.2

# Testing
pytest>=8.1.1
pytest-asyncio>=0.23.6
httpx>=0.27.0
